Завершить добавление функциональности - рекомендации
