+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -19,8 +19,6 @@ public interface FilmService {
 
     List<Film> getTopFilms(int maxCount);
 
-    List<Film> getTopFilms(int count, Integer genreId, Integer year);
-
     List<Film> getDirectorFilms(int directorId, String sortBy);
 
     void deleteFilmById(int filmId);
