+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/UserServiceImpl.java
@@ -1,25 +1,26 @@
 package ru.yandex.practicum.filmorate.service.impl;
 
+import jakarta.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.dal.FeedRepository;
-import ru.yandex.practicum.filmorate.dal.ReviewRepository;
+import ru.yandex.practicum.filmorate.dal.FilmRepository;
 import ru.yandex.practicum.filmorate.dal.UserRepository;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Feed;
+import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
-import java.util.List;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     @Qualifier("jdbcUserRepository")
     private final UserRepository userRepository;
-    private final FeedRepository feedRepository;
-    private final ReviewRepository reviewRepository;
+
+    private final FilmRepository filmRepository;
 
     @Override
     public User get(int id) {
@@ -51,7 +52,6 @@ public class UserServiceImpl implements UserService {
         User friend = userRepository.get(friendId)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id = " + friendId + " не найден"));
         userRepository.addFriend(user, friend);
-        addFriendFeed(userId, friendId, Feed.Operation.ADD);
     }
 
     @Override
@@ -61,7 +61,6 @@ public class UserServiceImpl implements UserService {
         User friend = userRepository.get(friendId)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id = " + friendId + " не найден"));
         userRepository.deleteFriend(user, friend);
-        addFriendFeed(userId, friendId, Feed.Operation.REMOVE);
     }
 
     @Override
@@ -81,30 +80,39 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<Feed> getFeed(int userId) {
-        User user = userRepository.get(userId)
-                .orElseThrow(() -> new NotFoundException("Пользователь с id = " + userId + " не найден"));
-        return feedRepository.findFeedByUserId(user.getId());
-    }
+    public List<Film> getRecommendedFilms(int userId) {
+        if (userId < 0) {
+            throw new ValidationException("Идентификатор пользователя должен быть положительным числом");
+        }
 
-    private void addFriendFeed(Integer userId, Integer friendId, Feed.Operation operation) {
-        feedRepository.create(new Feed(userId, friendId, Feed.EventType.FRIEND, operation));
-    }
+        // Фильмы, которые поставили лайк пользователь X, делавший запрос
+        List<Film> filmsLikedByUser = filmRepository.getFilmsLikedByUser(userId);
 
-    @Override
-    public void deleteUserById(int userId) {
-        User user = userRepository.get(userId)
-                .orElseThrow(() -> new NotFoundException("Пользователь с id = " + userId + " не найден"));
+        Set<Integer> filmsIdsLikedByUser = filmsLikedByUser.stream()
+                .map(Film::getId)
+                .collect(Collectors.toSet());
 
-        userRepository.deleteUserLikes(userId);
-        feedRepository.deleteUserFeed(userId);
-        userRepository.deleteUserFriends(userId);
-        reviewRepository.deleteUserReviewLikes(userId);
-        reviewRepository.deleteUserReviews(userId);
-        List<Integer> ids = reviewRepository.findUserReviewLikes(userId);
-        reviewRepository.deleteReviewLikesByUser(userId);
-        reviewRepository.refreshRatings(ids);
-        userRepository.deleteUserById(userId);
+        // Пользователи, которые поставили лайк те же самые фильмы, что и пользователь X
+        List<User> usersThatLikedSameFilms = userRepository.getUsersWithSameLikes(filmsIdsLikedByUser);
+        List<Integer> usersIds = usersThatLikedSameFilms.stream()
+                .map(User::getId)
+                .toList();
+        HashMap<Integer, List<Film>> foundUsersAllLikedFilms = filmRepository.getLikedFilmsByUsersIds(usersIds);
+
+
+        // Фильмы, которые не поставили лайк пользователь X, делавший запрос
+        List<Film> recommendedFilms = new ArrayList<>();
+
+        for (Map.Entry<Integer, List<Film>> entry : foundUsersAllLikedFilms.entrySet()) {
+            List<Film> likedByOtherUser = entry.getValue();
+
+            for (Film film : likedByOtherUser) {
+                if (!filmsLikedByUser.contains(film)) {
+                    recommendedFilms.add(film);
+                }
+            }
+        }
+        return recommendedFilms;
     }
 
 }
